rename run to pkg
