+++ cmd/nabu/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/gleanerio/nabu/run/cli"
+import "github.com/gleanerio/nabu/pkg/cli"
 
 func main() {
 	cli.Execute()
