+++ pkg/cli/object.go
@@ -2,34 +2,41 @@ package cli
 
 import (
 	"fmt"
-	"github.com/gleanerio/nabu/run"
+	"github.com/gleanerio/nabu/pkg"
 	"log"
 	"os"
 
 	"github.com/spf13/cobra"
 )
 
+var bucket, objectVal string
+
 // checkCmd represents the check command
-var tikaCmd = &cobra.Command{
-	Use:   "tika",
-	Short: "nabu tika command",
-	Long:  `(not implemented)This will read the configs/{cfgPath}/gleaner file, and try to connect to the minio server`,
+var objectCmd = &cobra.Command{
+	Use:   "objectVal",
+	Short: "nabu objectVal command",
+	Long:  `Load graph objectVal to triplestore`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("tike called")
-		err := run.Tika(viperVal, mc)
+		fmt.Println("objectVal called")
+		if objectVal == "" {
+
+		}
+		err := pkg.Object(viperVal, mc, bucketVal, objectVal)
 		if err != nil {
 			log.Fatal(err)
 			os.Exit(1)
 		}
 		os.Exit(0)
+
 	},
 }
 
 func init() {
-	rootCmd.AddCommand(tikaCmd)
+	rootCmd.AddCommand(objectCmd)
 
 	// Here you will define your flags and configuration settings.
-
+	// bucketVal is available at top level
+	objectCmd.Flags().StringVar(&objectVal, "objectVal", "", "objectVal to load")
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
 	// checkCmd.PersistentFlags().String("foo", "", "A help for foo")
