+++ pkg/cli/prune.go
