--- run/nabu.go
@@ -1,1 +1,0 @@
-package run
